api/requirements.txt
# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Image Processing
opencv-python-headless==4.9.0.80
numpy==1.26.4

# HTTP Client
httpx==0.27.0

# Validation & Models
pydantic==2.6.1
pydantic-settings==2.2.1

# Observability
structlog==24.1.0

# Utilities
python-dotenv==1.0.1

# Testing
pytest==8.0.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
